Fix JDK 21 compilation warnings (-Werror)
- Make setPropertyValue() final to eliminate 'this-escape' warning
  (calling overridable method in constructor)
- Use explicit cast in EventHeader compound assignment (int to byte)
- Use explicit cast in MessagePropertiesImpl compound assignment (long to int)

+++ bmq-sdk/src/main/java/com/bloomberg/bmq/impl/infr/proto/MessagePropertiesImpl.java
@@ -231,7 +231,7 @@ public class MessagePropertiesImpl implements MessageProperties {
         }
 
         // Add padding bytes
-        totalLength += numPaddingBytes;
+        totalLength += (int) numPaddingBytes;
 
         if (totalLength != propsAreaSize) {
             throw new IOException(
